// bga-cards.d.ts
type SideOrAngle = 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | 'top' | 'bottom' | 'left' | 'right'
// ...
interface CardStockSettings<T> {
	/**
	 * Indicate the card sorting (unset means no sorting, new cards will be added at the end).
	 * For example, use `sort: sortFunction('type', '-type_arg')` to sort by type then type_arg (in reversed order if prefixed with `-`).
	 * Be sure you typed the values correctly! Else '11' will be before '2'.
	 */
	sort?: SortFunction<T>
	/**
	 * The filter on card click event. Use setting from manager is unset.
	 */
	cardClickEventFilter?: CardClickEventFilter
	/**
	 * The style to apply to selectable cards. Use style from manager is unset.
	 */
	selectableCardStyle?: SelectionStyle
	/**
	 * The style to apply to selectable cards. Use style from manager is unset.
	 */
	unselectableCardStyle?: SelectionStyle
	/**
	 * The style to apply to selected cards. Use style from manager is unset.
	 */
	selectedCardStyle?: SelectionStyle
	/**
	 * The style to apply to the last played card. Default to class 'bga-cards_last-played-card'.
	 */
	lastPlayedCardStyle?: SelectionStyle
	/**
	 * Show a card counter on the stock. Not visible if unset.
	 */
	counter?: CardCounterSettings
	/**
	 * Say if a given card should be placed on this stock, based on card properties.
	 * For example, every card with location === 'hand' and location_arg == this.player_id should go to the current player hand.
	 * If unset, all cards on this stock must be added manually on this stock.
	 *
	 * @param card the card to place on a stock
	 * @returns true if the card should be placed on this stock.
	 */
	autoPlace?: (card: T) => boolean
}
interface AddCardSettings extends SlideAnimationSettings {
	/**
	 * The stock to take the card. It will automatically remove the card from the other stock.
	 */
	fromStock?: CardStock<any>
	/**
	 * The element to move the card from.
	 */
	fromElement?: HTMLElement
	/**
	 * Card side at the beginning of the animation. Default 'auto' to isCardVisible. Ignored if the card already exists.
	 */
	initialSide?: 'auto' | 'front' | 'back'
	/**
	 * Card side at the end of the animation. Default to initialSide.
	 */
	finalSide?: 'auto' | 'front' | 'back'
	forceToElement?: HTMLElement
	/**
	 * Force card position. Default to end of list. Do not use if sort is defined, as it will override it.
	 */
	index?: number
	/**
	 * Set if the card is selectable. Default is true, but will be ignored if the stock is not selectable.
	 */
	selectable?: boolean
	/**
	 * Indicates if we add a fade in effect when adding card (if it comes from an invisible or abstract element).
	 */
	fadeIn?: boolean
	/**
	 * For counters.
	 * Indicate if the card count is automatically updated when a card is added or removed. Default true.
	 */
	autoUpdateCardNumber?: boolean
}
interface RemoveCardSettings {
	slideTo?: HTMLElement
	fadeOut?: boolean
	/**
	 * For counters.
	 * Indicate if the card count is automatically updated when a card is added or removed. Default true.
	 */
	autoUpdateCardNumber?: boolean
}
// ...
interface DeckSettings<T> extends CardStockSettings<T> {
	/**
	 * Indicate the current top card.
	 */
	topCard?: T
	/**
	 * Indicate the current number of cards in the deck (default 0).
	 */
	cardNumber?: number
	/**
	 * Indicate if the card count is automatically updated when a card is added or removed.
	 */
	autoUpdateCardNumber?: boolean
	/**
	 * Indicate if the cards under the new top card must be removed (to forbid players to check the content of the deck with Inspect). Default true.
	 */
	autoRemovePreviousCards?: boolean
	/**
	 * Indicate the thresholds to add 1px to the thickness of the pile. Default [0, 2, 5, 10, 20, 30].
	 */
	thicknesses?: number[]
	/**
	 * Shadow direction. Default 'bottom-right'.
	 */
	shadowDirection?: SideOrAngle
	/**
	 * A generator of fake cards, to generate decks top card automatically.
	 * Default is manager `fakeCardGenerator` method.
	 *
	 * @param deckId the deck id
	 * @return the fake card to be generated (usually, only informations to show back side)
	 */
	fakeCardGenerator?: (deckId: string) => T
}
interface AddCardToDeckSettings extends AddCardSettings {
	/**
	 * Indicate if the cards under the new top card must be removed (to forbid players to check the content of the deck with Inspect). Default true.
	 */
	autoRemovePreviousCards?: boolean
}
interface RemoveCardFromDeckSettings extends RemoveCardSettings {
	/**
	 * Indicate if the card count is automatically updated when a card is added or removed.
	 */
	autoUpdateCardNumber?: boolean
}
interface ShuffleAnimationSettings<T> {
	/**
	 * Number of cards used for the animation (will use cardNumber is inferior to this number).
	 * Default: 10.
	 */
	animatedCardsMax?: number
	/**
	 * Card generator for the animated card. Should only show the back of the cards.
	 * Default if fakeCardGenerator from Deck (or Manager if unset in Deck).
	 */
	fakeCardSetter?: (card: T, index: number) => void
	/**
	 * Time to wait after shuffle, in case it is chained with other animations, to let the time to understand it's 2 different animations.
	 * Default is 500ms.
	 */
	pauseDelayAfterAnimation?: number
}
/**
 * Stock to represent a deck. (pile of cards, with a fake 3d effect of thickness).
 */
declare class Deck<T> extends CardStock<T> {
	protected manager: CardManager<T>
	protected element: HTMLElement
	protected cardNumber: number
	protected autoUpdateCardNumber: boolean
	protected autoRemovePreviousCards: boolean
	protected fakeCardGenerator?: (deckId: string) => T
	protected thicknesses: number[]
	constructor(manager: CardManager<T>, element: HTMLElement, settings: DeckSettings<T>)
	/**
	 * Get the the cards number.
	 *
	 * @returns the cards number
	 */
	getCardNumber(): number
	/**
	 * Set the the cards number.
	 *
	 * @param cardNumber the cards number
	 * @param topCard the deck top card. If unset, will generated a fake card (default). Set it to null to not generate a new topCard.
	 */
	setCardNumber(cardNumber: number, topCard?: T | null | undefined): Promise<boolean>
	addCard(card: T, settings?: AddCardToDeckSettings): Promise<boolean>
	cardRemoved(card: T, settings?: RemoveCardFromDeckSettings): void
	removeAll(settings?: RemoveCardFromDeckSettings): Promise<boolean>
	getTopCard(): T | null
	/**
	 * Shows a shuffle animation on the deck
	 *
	 * @param settings a `ShuffleAnimationSettings` object
	 * @returns promise when animation ends
	 */
	shuffle(settings?: ShuffleAnimationSettings<T>): Promise<boolean>
	protected getFakeCard(): T
	/**
	 * Returns the card count in the deck (what the player think there is, for decks, the real number of cards for all visible card stocks).
	 *
	 * @returns the number of card in the stock
	 */
	getCardCount(): number
}
// ...
interface LineStockSettings<T> extends CardStockSettings<T> {
	/**
	 * Indicate if the line should wrap when needed (default wrap)
	 */
	wrap?: 'wrap' | 'nowrap'
	/**
	 * Indicate the line direction (default row)
	 */
	direction?: 'row' | 'column'
	/**
	 * indicate if the line should be centered (default yes)
	 */
	center?: boolean
	/**
	 * CSS to set the gap between cards. '8px' if unset.
	 */
	gap?: string
}
/**
 * A basic stock for a list of cards, based on flex.
 */
declare class LineStock<T> extends CardStock<T> {
	protected manager: CardManager<T>
	protected element: HTMLElement
	/**
	 * @param manager the card manager
	 * @param element the stock element (should be an empty HTML Element)
	 * @param settings a `LineStockSettings` object
	 */
	constructor(manager: CardManager<T>, element: HTMLElement, settings?: LineStockSettings<T>)
}